requirements.txt
fastapi
uvicorn
pymupdf
sentence-transformers
faiss-cpu
groq
python-multipart
python-dotenv
numpy
